chore: Move misc file's contents to mod.rs in tests dir.
