+++ core/src/test/mod.rs
@@ -1,7 +1,6 @@
 pub mod common;
 mod deposit_and_withdraw_e2e;
 mod full_flow;
-mod misc;
 mod musig2;
 mod state_manager;
 mod taproot;
